set a new owner of a tweet/message
Check if the address of the new owner is not 0x0.
Check if the message with that particular Id has already been created.

Set only the owner field of the struct.

When using 'memory', a copy of the struct is made in memory and then updated. On the other hand, when using 'storage', a reference to the struct in storage is obtained, so any modifications will be persisted.