add a function to remove a message from an account
When transferring the ownership of a tweet, the message is added to the new owner's array and removed from the current owner's array of owned tweets.

The array of owned tweets stores the IDs of the tweets owned by the account.
- Copy the last element of the array (last ID) to the position where the item to be removed is in
- Update the information about where this copied ID is located (before it was the last element of the array, now it is in the position of the removed ID)
- Pop the last element of the array.

The function Pop() removes the last item of the array and reduces its length. As the last element has already been copied, it just moves to another position inside the array.