add custom font
Add the Futura font face to the app
- Futura regular
- Futura italic
- Futura heavy

The source of the fonts are located in ./fonts/ folder and the fonts are imported by 'pubspec.yaml'. The Yaml file defines the name of the family and the path to the font files.

After importing by pubspec.yaml, the fonts can be used in fontFamily property of TextStyle.

Signed-off-by: Fabio Benjovengo <fabio.benjovengo@gmail.com>