create smart contract to manage the user account sign-in data
Create smart contract called AccountRegistrationInfo that implements the functions and state variables to store the sign-in information of user accounts on the blockchain. The contract includes the function addUpdateAccount which creates or updates the sign-in information for a user on the blockchain.

'accountsSignInData': mapping(address => AccountData)
- The sign-in information for a user is stored in a mapping from the address of the user's account to the struct containing the user's info.

'accountsModifications': mapping(address => array of IDs)
 - For each struct that is stored, a unique ID is generated to uniquely identify the data. Whenever there are changes in the information, a new ID is generated and added to an array that is referenced in a mapping that uses the user's address as the key. The most recent data corresponds to the ID stored in the last entry of the array.

Signed-off-by: Fabio Benjovengo <fabio.benjovengo@gmail.com>